FAC-651:feature:Merge UpdateAttributesOrderUseCaseParamTests
